+++ g_spawn.c
@@ -348,11 +348,11 @@ ED_NewString
 char *ED_NewString (char *string)
 {
 	char	*newb, *new_p;
-	int		i,l;
+	size_t	i,l;
 	
 	l = strlen(string) + 1;
 
-	newb = gi.TagMalloc (l, TAG_LEVEL);
+	newb = gi.TagMalloc ((int)l, TAG_LEVEL);
 
 	new_p = newb;
