+++ m_medic.c
@@ -628,7 +628,7 @@ void medic_cable_attack (edict_t *self)
 
 	gi.WriteByte (svc_temp_entity);
 	gi.WriteByte (TE_MEDIC_CABLE_ATTACK);
-	gi.WriteShort (self - g_edicts);
+	gi.WriteShort ((int)(self - g_edicts));
 	gi.WritePosition (start);
 	gi.WritePosition (end);
 	gi.multicast (self->s.origin, MULTICAST_PVS);
