+++ p_client.c
@@ -867,7 +867,7 @@ edict_t *SelectCoopSpawnPoint (edict_t *ent)
 	edict_t	*spot = NULL;
 	char	*target;
 
-	index = ent->client - game.clients;
+	index = (int)(ent->client - game.clients);
 
 	// player 0 starts in normal player spawn point
 	if (!index)
@@ -990,7 +990,7 @@ void CopyToBodyQue (edict_t *ent)
 
 	gi.unlinkentity (body);
 	body->s = ent->s;
-	body->s.number = body - g_edicts;
+	body->s.number = (int)(body - g_edicts);
 
 	body->svflags = ent->svflags;
 	VectorCopy (ent->mins, body->mins);
@@ -1099,7 +1099,7 @@ void spectator_respawn (edict_t *ent)
 	if (!ent->client->pers.spectator)  {
 		// send effect
 		gi.WriteByte (svc_muzzleflash);
-		gi.WriteShort (ent-g_edicts);
+		gi.WriteShort ((int)(ent-g_edicts));
 		gi.WriteByte (MZ_LOGIN);
 		gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1131,7 +1131,7 @@ void PutClientInServer (edict_t *ent)
 {
 	vec3_t	mins = {-16, -16, -24};
 	vec3_t	maxs = {16, 16, 32};
-	int		index;
+	size_t		index;
 	vec3_t	spawn_origin, spawn_angles;
 	gclient_t	*client;
 	int		i;
@@ -1241,7 +1241,7 @@ void PutClientInServer (edict_t *ent)
 
 	// clear entity state values
 	ent->s.effects = 0;
-	ent->s.skinnum = ent - g_edicts - 1;
+	ent->s.skinnum = (int)(ent - g_edicts - 1);
 	ent->s.modelindex = 255;		// will use the skin specified model
 	ent->s.modelindex2 = 255;		// custom gun model
 	ent->s.frame = 0;
@@ -1313,7 +1313,7 @@ void ClientBeginDeathmatch (edict_t *ent)
 	{
 		// send effect
 		gi.WriteByte (svc_muzzleflash);
-		gi.WriteShort (ent-g_edicts);
+		gi.WriteShort ((int)(ent-g_edicts));
 		gi.WriteByte (MZ_LOGIN);
 		gi.multicast (ent->s.origin, MULTICAST_PVS);
 	}
@@ -1377,7 +1377,7 @@ void ClientBegin (edict_t *ent)
 		if (game.maxclients > 1)
 		{
 			gi.WriteByte (svc_muzzleflash);
-			gi.WriteShort (ent-g_edicts);
+			gi.WriteShort ((int)(ent-g_edicts));
 			gi.WriteByte (MZ_LOGIN);
 			gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1426,7 +1426,7 @@ void ClientUserinfoChanged (edict_t *ent, char *userinfo)
 	// set skin
 	s = Info_ValueForKey (userinfo, "skin");
 
-	playernum = ent-g_edicts-1;
+	playernum = (int)(ent-g_edicts-1);
 
 	// combine name and skin into a configstring
 	gi.configstring (CS_PLAYERSKINS+playernum, va("%s\\%s", ent->client->pers.netname, s) );
@@ -1554,7 +1554,7 @@ void ClientDisconnect (edict_t *ent)
 
 	// send effect
 	gi.WriteByte (svc_muzzleflash);
-	gi.WriteShort (ent-g_edicts);
+	gi.WriteShort ((int)(ent-g_edicts));
 	gi.WriteByte (MZ_LOGOUT);
 	gi.multicast (ent->s.origin, MULTICAST_PVS);
 
@@ -1565,7 +1565,7 @@ void ClientDisconnect (edict_t *ent)
 	ent->classname = "disconnected";
 	ent->client->pers.connected = false;
 
-	playernum = ent-g_edicts-1;
+	playernum = (int)(ent-g_edicts-1);
 	gi.configstring (CS_PLAYERSKINS+playernum, "");
 }
